+++ src/client/index.js
@@ -4,6 +4,9 @@ import { BrowserRouter as Router } from 'react-router-dom';
 import App from './App';
 import { configure } from '@store';
 import { Provider } from 'react-redux'
+import jwt_decode from 'jwt-decode';
+import setAuthToken from './utils/setAuthToken';
+import { setCurrentUser, logoutUser } from '@store/actions/authActions';
 
 import './styles/index.scss';
 
@@ -12,6 +15,22 @@ window.addEventListener('DOMContentLoaded', () => {
     delete window.__STATE__;
 
     const store = configure(preloadedState);
+
+    if (localStorage.jwtToken) {
+        setAuthToken(localStorage.jwtToken);
+
+        const decoded = jwt_decode(localStorage.jwtToken);
+
+        store.dispatch(setCurrentUser(decoded));
+
+        const currentTime = Date.now() / 1000;
+
+        if (decoded.exp < currentTime) {
+            store.dispatch(logoutUser());
+
+            window.location.href = '/login';
+        }
+    }
 
     ReactDOM.render((<Provider store={store}>
         <Router>
