+++ src/client/store/index.js
@@ -3,8 +3,6 @@ import thunk from 'redux-thunk';
 
 import authReducer from './reducers/authReducer';
 import errorReducer from './reducers/errorReducer';
-import profileReducer from './reducers/profileReducer';
-import postReducer from './reducers/postReducer';
 import galleryReducer from './reducers/galleryReducer';
 
 // const store = createStore(rootReducer, initialState,
@@ -17,14 +15,13 @@ export const configure = (preloadedState = {}) => {
     const rootReducer = combineReducers({
         auth: authReducer,
         errors: errorReducer,
-        profile: profileReducer,
-        post: postReducer,
         gallery: galleryReducer
     });
 
     return createStore(
         rootReducer,
         preloadedState,
-        compose(applyMiddleware(...middleware))
+        compose(applyMiddleware(...middleware)),
+        window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__ ()
     );
 }
